refactor: move decoder to layers
