+++ mmdet/models/detectors/dino.py
@@ -1,5 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-import math
 from typing import Dict, List, Tuple, Union
 
 import torch
@@ -10,10 +9,9 @@ from mmdet.registry import MODELS
 from mmdet.structures import OptSampleList, SampleList
 from mmdet.structures.bbox import bbox_xyxy_to_cxcywh
 from mmdet.utils import OptConfigType
-from ..layers import MLP, SinePositionalEncoding, inverse_sigmoid
-from .deformable_detr import (DeformableDETR, DeformableDetrTransformerDecoder,
-                              DeformableDetrTransformerEncoder,
-                              MultiScaleDeformableAttention)
+from ..layers import (DeformableDetrTransformerEncoder, DinoTransformerDecoder,
+                      SinePositionalEncoding, inverse_sigmoid)
+from .deformable_detr import DeformableDETR, MultiScaleDeformableAttention
 
 
 @MODELS.register_module()
@@ -358,111 +356,6 @@ class DINO(DeformableDETR):
         decoder_outputs_dict = dict(
             hidden_states=inter_states, references=references)
         return decoder_outputs_dict
-
-
-class DinoTransformerDecoder(DeformableDetrTransformerDecoder):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._init_layers()
-
-    def _init_layers(self):
-        super()._init_layers()
-        self.ref_point_head = MLP(self.embed_dims * 2, self.embed_dims,
-                                  self.embed_dims, 2)
-        self.norm = nn.LayerNorm(self.embed_dims)
-
-    @staticmethod
-    def gen_sineembed_for_position(pos_tensor):
-        # n_query, bs, _ = pos_tensor.size()
-        # sineembed_tensor = torch.zeros(n_query, bs, 256)
-        scale = 2 * math.pi
-        dim_t = torch.arange(
-            128, dtype=torch.float32, device=pos_tensor.device)
-        dim_t = 10000**(2 * (dim_t // 2) / 128)
-        x_embed = pos_tensor[:, :, 0] * scale
-        y_embed = pos_tensor[:, :, 1] * scale
-        pos_x = x_embed[:, :, None] / dim_t
-        pos_y = y_embed[:, :, None] / dim_t
-        pos_x = torch.stack((pos_x[:, :, 0::2].sin(), pos_x[:, :, 1::2].cos()),
-                            dim=3).flatten(2)
-        pos_y = torch.stack((pos_y[:, :, 0::2].sin(), pos_y[:, :, 1::2].cos()),
-                            dim=3).flatten(2)
-        if pos_tensor.size(-1) == 2:
-            pos = torch.cat((pos_y, pos_x), dim=2)
-        elif pos_tensor.size(-1) == 4:
-            w_embed = pos_tensor[:, :, 2] * scale
-            pos_w = w_embed[:, :, None] / dim_t
-            pos_w = torch.stack(
-                (pos_w[:, :, 0::2].sin(), pos_w[:, :, 1::2].cos()),
-                dim=3).flatten(2)
-
-            h_embed = pos_tensor[:, :, 3] * scale
-            pos_h = h_embed[:, :, None] / dim_t
-            pos_h = torch.stack(
-                (pos_h[:, :, 0::2].sin(), pos_h[:, :, 1::2].cos()),
-                dim=3).flatten(2)
-
-            pos = torch.cat((pos_y, pos_x, pos_w, pos_h), dim=2)
-        else:
-            raise ValueError('Unknown pos_tensor shape(-1):{}'.format(
-                pos_tensor.size(-1)))
-        return pos
-
-    def forward(self,
-                query,
-                *args,
-                reference_points=None,
-                valid_ratios=None,
-                reg_branches=None,
-                **kwargs):
-        output = query
-        intermediate = []
-        intermediate_reference_points = [reference_points]
-        for lid, layer in enumerate(self.layers):
-            if reference_points.shape[-1] == 4:
-                reference_points_input = \
-                    reference_points[:, :, None] * torch.cat(
-                        [valid_ratios, valid_ratios], -1)[:, None]
-            else:
-                assert reference_points.shape[-1] == 2
-                reference_points_input = \
-                    reference_points[:, :, None] * valid_ratios[:, None]
-
-            query_sine_embed = self.gen_sineembed_for_position(
-                reference_points_input[:, :, 0, :])
-            query_pos = self.ref_point_head(query_sine_embed)
-
-            query_pos = query_pos.permute(1, 0, 2)
-            output = layer(
-                output,
-                *args,
-                query_pos=query_pos,
-                reference_points=reference_points_input,
-                **kwargs)
-            output = output.permute(1, 0, 2)
-
-            if reg_branches is not None:
-                tmp = reg_branches[lid](output)
-                assert reference_points.shape[-1] == 4
-                # TODO: should do earlier
-                new_reference_points = tmp + inverse_sigmoid(
-                    reference_points, eps=1e-3)
-                new_reference_points = new_reference_points.sigmoid()
-                reference_points = new_reference_points.detach()
-
-            output = output.permute(1, 0, 2)
-            if self.return_intermediate:
-                intermediate.append(self.norm(output))
-                intermediate_reference_points.append(new_reference_points)
-                # NOTE this is for the "Look Forward Twice" module,
-                # in the DeformDETR, reference_points was appended.
-
-        if self.return_intermediate:
-            return torch.stack(intermediate), torch.stack(
-                intermediate_reference_points)
-
-        return output, reference_points
 
 
 class CdnQueryGenerator:
